[PHPStanStaticTypeMapper] Decouple new services to handle PHPStan type mapping
