+++ packages/PHPStanStaticTypeMapper/config/config.yaml
@@ -1,0 +1,8 @@
+services:
+    _defaults:
+        public: true
+        autowire: true
+
+    Rector\PHPStanStaticTypeMapper\:
+        resource: '../src'
+        exclude:
